Tweaked the unauthorized logic. A user should only see the unauthorized screen if they had a valid login, but are not a hubmap user.
if they have an old or invalid token (which should only happen to
returning users, not people logging in), they should just be met with
the regular login screen.

The metadata/usergroups endpoint of ingest-api appears to reject if users
are not part of hubmap, so all I'm checking for now is if that endpoint
returns a 401. If the endpoint returns a 401 the regular login screen
appears. If it returns its 403 or returns a 200 but hubmap-read is not in
user groups, the unauthorized screen appears.